feat: implementar insights controla o contexto e os componentes da interface do usuário para seleção do escopo do relatório
