+++ apps/web/components/user/pages/feedbacksInsightsReport/ui.types.ts
@@ -19,6 +19,34 @@ export interface InsightsReportHeaderSectionProps {
 }
 
 /**
+ * Props dos controles de insights no header global.
+ * Usado em: components/user/pages/feedbacksInsightsReport/InsightsHeaderControls.tsx.
+ */
+export interface InsightsHeaderControlsProps {
+  refreshing: boolean;
+  analyzingRaw: boolean;
+  canAnalyze: boolean;
+  availableScopes: InsightScopeOption[];
+  selectedScope: InsightScopeOption;
+  selectedCatalogItemId: string;
+  catalogItemOptions: InsightsCatalogItemOption[];
+  onScopeChange: (scope: InsightScopeOption) => void;
+  onCatalogItemChange: (id: string) => void;
+  onRefreshSelected: () => void;
+  onAnalyzeRaw: () => void;
+}
+
+/**
+ * Props do seletor radial de escopos.
+ * Usado em: components/user/pages/feedbacksInsightsReport/ScopeSelectorRadial.tsx.
+ */
+export interface ScopeSelectorRadialProps {
+  options: InsightScopeOption[];
+  selected: InsightScopeOption;
+  onChange: (scope: InsightScopeOption) => void;
+}
+
+/**
  * Tons de sentimento usados no bloco de humor.
  * Usado em: components/user/pages/feedbacksInsightsReport/InsightsReportMoodSection.tsx.
  */
